+++ dotextensions/server/handlers/__init__.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger('handler')
