+++ dotextensions/server/postman2_1/__init__.py
@@ -5,6 +5,8 @@
 # and then, to convert JSON from a string, do
 #
 #     result = postman21_collection_from_dict(json.loads(json_string))
+from ..utils import from_str, from_none, from_union, from_list, to_class, to_enum, from_bool, \
+    from_int, from_dict, from_float, to_float
 
 POSTMAN_2_1 = 'https://schema.getpostman.com/json/collection/v2.1.0/collection.json'
 # To use this code, make sure you
@@ -16,69 +18,7 @@ POSTMAN_2_1 = 'https://schema.getpostman.com/json/collection/v2.1.0/collection.j
 #     result = postman_collection21_from_dict(json.loads(json_string))
 
 from enum import Enum
-from typing import Optional, Any, List, Union, Dict, TypeVar, Callable, Type, cast
-
-T = TypeVar("T")
-EnumT = TypeVar("EnumT", bound=Enum)
-
-
-def from_str(x: Any) -> str:
-    assert isinstance(x, str)
-    return x
-
-
-def from_none(x: Any) -> Any:
-    assert x is None
-    return x
-
-
-def from_union(fs, x):
-    for f in fs:
-        try:
-            return f(x)
-        except:
-            pass
-    assert False
-
-
-def from_list(f: Callable[[Any], T], x: Any) -> List[T]:
-    assert isinstance(x, list)
-    return [f(y) for y in x]
-
-
-def to_class(c: Type[T], x: Any) -> dict:
-    assert isinstance(x, c)
-    return cast(Any, x).to_dict()
-
-
-def to_enum(c: Type[EnumT], x: Any) -> EnumT:
-    assert isinstance(x, c)
-    return x.value
-
-
-def from_bool(x: Any) -> bool:
-    assert isinstance(x, bool)
-    return x
-
-
-def from_int(x: Any) -> int:
-    assert isinstance(x, int) and not isinstance(x, bool)
-    return x
-
-
-def from_dict(f: Callable[[Any], T], x: Any) -> Dict[str, T]:
-    assert isinstance(x, dict)
-    return {k: f(v) for (k, v) in x.items()}
-
-
-def from_float(x: Any) -> float:
-    assert isinstance(x, (float, int)) and not isinstance(x, bool)
-    return float(x)
-
-
-def to_float(x: Any) -> float:
-    assert isinstance(x, float)
-    return x
+from typing import Optional, Any, List, Union, Dict
 
 
 class ApikeyElement:
